fetch/api/response/response-consume-empty.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Response consume empty bodies</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#response">
    <meta name="help" href="https://fetch.spec.whatwg.org/#body-mixin">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    function checkBodyText(test, response) {
      return response.text().then(function(bodyAsText) {
        assert_equals(bodyAsText, "", "Resolved value should be empty");
      });
    }

    function checkBodyBlob(test, response) {
      return response.blob().then(function(bodyAsBlob) {
        var promise = new Promise(function(resolve, reject) {
          var reader = new FileReader();
          reader.onload = function(evt) {
            resolve(reader.result)
          };
          reader.onerror = function() {
            reject("Blob's reader failed");
          };
          reader.readAsText(bodyAsBlob);
        });
        return promise.then(function(body) {
          assert_equals(body, "", "Resolved value should be empty");
        });
      });
    }

    function checkBodyArrayBuffer(test, response) {
      return response.arrayBuffer().then(function(bodyAsArrayBuffer) {
        assert_equals(bodyAsArrayBuffer.byteLength, 0, "Resolved value should be empty");
      });
    }

    function checkBodyJSON(test, response) {
      return response.json().then(
        function(bodyAsJSON) {
          assert_unreached("JSON parsing should fail");
        },
        function() {
        });
    }

    function checkBodyFormData(test, response) {
      return response.formData().then(function(bodyAsFormData) {
        assert_true(bodyAsFormData instanceof FormData, "Should receive a FormData");
     });
    }

    function checkBodyFormDataError(test, response) {
      return promise_rejects(test, new TypeError(), response.formData());
    }

    function checkBodyFormDataText(test, response) {
      return response.text().then(function(bodyAsText) {
        assert_true(response.headers.has("Content-Type"), "Response contains Content-Type header");
        var boundaryMatches = response.headers.get("Content-Type").match(/;\s*boundary=("?)([^";\s]*)\1/);
        assert_true(!!boundaryMatches, "Response contains boundary parameter");
        var boundary = boundaryMatches[2];
        assert_equals(bodyAsText, "--" + boundary + "--\r\n");
     });
    }

    function checkResponseWithNoBody(bodyType, checkFunction, headers = []) {
      promise_test(function(test) {
        var response = new Response(undefined, { "headers": headers });
        assert_false(response.bodyUsed);
        return checkFunction(test, response).then(function() {
          assert_false(response.bodyUsed);
        });
      }, "Consume response's body as " + bodyType);
    }

    checkResponseWithNoBody("text", checkBodyText);
    checkResponseWithNoBody("blob", checkBodyBlob);
    checkResponseWithNoBody("arrayBuffer", checkBodyArrayBuffer);
    checkResponseWithNoBody("json (error case)", checkBodyJSON);
    checkResponseWithNoBody("formData with correct multipart type (error case)", checkBodyFormDataError, [["Content-Type", 'multipart/form-data; boundary="boundary"']]);
    checkResponseWithNoBody("formData with correct urlencoded type", checkBodyFormData, [["Content-Type", "application/x-www-form-urlencoded;charset=UTF-8"]]);
    checkResponseWithNoBody("formData without correct type (error case)", checkBodyFormDataError);

    function checkResponseWithEmptyBody(bodyType, body, checkFunction) {
      promise_test(function(test) {
        var response = new Response(body);
        assert_false(response.bodyUsed, "bodyUsed is false at init");
        return checkFunction(test, response).then(function() {
          assert_true(response.bodyUsed, "bodyUsed is true after being consumed");
        });
      }, "Consume empty " + bodyType + " response body as " + (checkFunction == checkBodyArrayBuffer ? "arrayBuffer" : "text"));
    }

    checkResponseWithEmptyBody("blob", new Blob([], { "type" : "text/plain" }), checkBodyArrayBuffer);
    checkResponseWithEmptyBody("text", "", checkBodyArrayBuffer);
    checkResponseWithEmptyBody("blob", new Blob([], { "type" : "text/plain" }), checkBodyText);
    checkResponseWithEmptyBody("text", "", checkBodyText);
    checkResponseWithEmptyBody("URLSearchParams", new URLSearchParams(""), checkBodyText);
    checkResponseWithEmptyBody("FormData", new FormData(), checkBodyFormDataText);
    checkResponseWithEmptyBody("ArrayBuffer", new ArrayBuffer(), checkBodyText);
    </script>
  </body>
</html>
